Add dev preview. Fix whitespace issue.
Now you can use the compile-and-watch dev preview by running `npm start`.

There was an issue where strings ending in whitespace where not
measured correctly because an SVG text element with "AA" has
the same bounding box as "AA    ". To circumvent this issue,
we use the built-in guard strings when the measured text
ends in whitespace.